requirements.txt
llama-index==0.10.67.post1
llama-index-llms-ollama==0.1.5
llama-index-embeddings-huggingface==0.2.2
llama-index-vector-stores-chroma==0.1.10
chromadb==0.5.5
streamlit==1.38.0
python-dotenv==1.0.1
pymupdf==1.24.9
pillow==10.4.0
sentence-transformers==3.0.1
torch==2.4.1